Remove logo (#4)
We don't need and want this for the ACB and NFB conferences. These are blind or partially blind people, they don't care about the logo section.

It was, however, confusing them, because it was representing itself as something they should care about. When we get back from the conference, we can work on improving this element and re-introduce it.